perf/x86/intel: Add attr_update for Hybrid PMUs
Bugzilla: https://bugzilla.redhat.com/1838573

upstream
========
commit 58ae30c29a370c09eb49e0007d881a9aed13c5a3
Author: Kan Liang <kan.liang@linux.intel.com>
Date: Mon Apr 12 07:30:58 2021 -0700

description
===========
The attribute_group for Hybrid PMUs should be different from the
previous
cpu PMU. For example, cpumask is required for a Hybrid PMU. The PMU type
should be included in the event and format attribute.

Add hybrid_attr_update for the Hybrid PMU.
Check the PMU type in is_visible() function. Only display the event or
format for the matched Hybrid PMU.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>