perf/x86/intel: Fix fixed counter check warning for some Alder Lake
Bugzilla: https://bugzilla.redhat.com/1838573

upstream
========
commit ee72a94ea4a6d8fa304a506859cd07ecdc0cf5c4
Author: Kan Liang <kan.liang@linux.intel.com>
Date:   Fri Jun 18 08:12:52 2021 -0700

description
===========
For some Alder Lake machine, the below fixed counter check warning may be
triggered.

[    2.010766] hw perf events fixed 5 > max(4), clipping!

Current perf unconditionally increases the number of the GP counters and
the fixed counters for a big core PMU on an Alder Lake system, because
the number enumerated in the CPUID only reflects the common counters.
The big core may has more counters. However, Alder Lake may have an
alternative configuration. With that configuration,
the X86_FEATURE_HYBRID_CPU is not set. The number of the GP counters and
fixed counters enumerated in the CPUID is accurate. Perf mistakenly
increases the number of counters. The warning is triggered.

Directly use the enumerated value on the system with the alternative
configuration.

Fixes: f83d2f91d259 ("perf/x86/intel: Add Alder Lake Hybrid support")
Link: https://lore.kernel.org/r/1624029174-122219-2-git-send-email-kan.liang@linux.intel.com
Signed-off-by: Michael Petlan <mpetlan@redhat.com>