tasklist_lock, ppc64: Provide infrastructure to support specific use of qrwlock
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1978377
Upstream Status: RHEL only

On large ppcle64 SMP systems with many CPUs, the use of unfair rwlock
for tasklist_lock can sometimes lead to soft or hard lockups. To avoid
those problems, we have to switch to use qrwlock which is a fair rwlock.
Unfortunately, enabling qspinlock or qrwlock for the whole arch will
break kABI. As a result, we aim for the conversion of the internal
tasklist_lock to qrwlock only, the other rwlocks will still use the
unfair version.

This patch provides the infrastructure that enables the use of qrwlock
for tasklist_lock. For kernel files that use only use tasklist_lock
without other rwlocks, including <linux/rh_tasklist_lock.h> is enough
for the qrwlock conversion. For other kernel files that use tasklist_lock
and other rwlocks, the locking functions for tasklist_lock have to be
modified to use the qrwlock equivalent.

Note that the unfair ppcle rwlock is 4 bytes in size whereas
qrwlock needs 8 bytes. So we have to define qrwlock_t in
arch/powerpc/include/asm/spinlock.h.

The next patch will perform the actual conversion.

Signed-off-by: Waiman Long <longman@redhat.com>