Revert "ice: Add initial support framework for LAG"
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1999016
Upstream Status: RHEL-only

This reverts commit 19c5dbb4a23cbcf57f524def2fbfae72c228fa3f.
There's something wrong in the commit about refcounting of ice netdevs
when put into bonding. The problem is visible when rebooting. There are
repeated error messages:
  unregister_netdevice: waiting for ens1f1 to become free. Usage count = 1

Intel is looking into it, but no proper fix is available yet. Let's
revert the LAG framework patch. The HW offload for bonding that the
framework is meant to be a basis for is not yet implemented anyway.

Conflicts:
	drivers/net/ethernet/intel/ice/ice.h
More fields were added to struct ice_pf since the commit we're
reverting. In this revert, we only need to remove the lag pointer.

	drivers/net/ethernet/intel/ice/ice_main.c
Context changed because of the addition of the PTP feature.

Signed-off-by: Michal Schmidt <mschmidt@redhat.com>