cgroup/debug: Fix lockdep splat with "%pK" format specifier
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2007919
Upstream Status: RHEL only

The BZ reported the following lockdep splat:

[ 1732.553607] =====================================================
[ 1732.560413] WARNING: SOFTIRQ-safe -> SOFTIRQ-unsafe lock order detected
[ 1732.567801] 4.18.0-345.el8.x86_64+debug #1 Not tainted
[ 1732.573536] -----------------------------------------------------
    :
[ 1732.776898]  Possible interrupt unsafe locking scenario:

[ 1732.784469]        CPU0                    CPU1
[ 1732.789517]        ----                    ----
[ 1732.794565]   lock(&selinux_ss.policy_rwlock);
[ 1732.799528]                                local_irq_disable();
[ 1732.806137]                                lock(css_set_lock);
[ 1732.812648]                                lock(&selinux_ss.policy_rwlock);
[ 1732.820413]   <Interrupt>
[ 1732.823335]     lock(css_set_lock);
[ 1732.827222]
                *** DEADLOCK ***

The lockdep splat is caused by the fact that the debug controller use the
"%pK" format specifier to print out address of cset's with css_set_lock
held. Under some circumstances, the use of "%pK" format specifier may
acquire the selinux_ss.policy_rwlock.

Upstream had get rid of selinux_ss.policy_rwlock since v5.10 and used RCU
instead. So upstream kernel shouldn't have this problem.  It will also
be a lot of work if we have to backport the relevant selinux commits.

On the other hand, the debug controller involved in this lockdep splat
is a debug kernel only controller and its purpose is just for debugging
by developers.  This controller had not been changed for quite a while
and it is not expecting further change anyway.  It is easier to just
do a simple RHEL-only patch to address this problem.

To avoid this possible deadlock scenario, we have to abandon the use of
the "%pK" format specifier and just use "%p" to always hash the cset
addresses. The actual cset addresses aren't that important as long as
they are unique for matching purpose.

Signed-off-by: Waiman Long <longman@redhat.com>