Revert "Merge: mt7921e: enable new Mediatek wireless hardware"
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2009501
Upstream-status: RHEL only

This reverts the commits made by MR 1341, which introduced the support for
the NIC Wifi 6 MT7921 and failed final verification. Since the card is
unusable we need to remove it from the tree and move the enablement to the
next release.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>