scsi: iscsi: Get ref to conn during reset handling
The comment in iscsi_eh_session_reset is wrong and we don't wait for the
EH to complete before tearing down the conn. This has us get a ref to the
conn when we are not holding the eh_mutex/frwd_lock so it does not get
freed from under us.

Link: https://lore.kernel.org/r/20210525181821.7617-13-michael.christie@oracle.com
Reviewed-by: Lee Duncan <lduncan@suse.com>
Signed-off-by: Mike Christie <michael.christie@oracle.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit fda290c5ae98b591797c96e55a2cfbd30d46a706)
Signed-off-by: Chris Leech <cleech@redhat.com>
Bugzilla: http://bugzilla.redhat.com/2011470