tools: bpf: add SO_BINDTOIFINDEX to socket.h in tools
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1947006

Upstream: partial revert of a RHEL fix, restoring upstream state

This is a partial revert of RHEL 8 commit cb9a52e7bce8 ("[tools] bpf: Sync
RHEL version of asm-generic/socket.h to tools/"). That commit synced the
tools/include/uapi/asm-generic/socket.h file to match the RHEL 8 version
that was current at that time.

While that was the correct thing to do, the result were two changes: removal
of the y38 socket options that are not avaliable in RHEL 8, and removal of
SO_BINDTOIFINDEX. SO_BINDTOIFINDEX was not present in RHEL 8 at that time
but it is present now.

Conflicts: This commit reintroduces SO_BINDTOIFINDEX to
tools/include/uapi/asm-generic/socket.h. As such, it is a partial rebackport
of:

commit 13a748ea6df1fee13fdcbd9d10a3d70f9e188b48
Author: Martin KaFai Lau <kafai@fb.com>
Date:   Thu Jun 13 15:00:03 2019 -0700

    bpf: Sync asm-generic/socket.h to tools/

    SO_DETACH_REUSEPORT_BPF is needed for the test in the next patch.
    It is defined in the socket.h.

    Signed-off-by: Martin KaFai Lau <kafai@fb.com>
    Reviewed-by: Stanislav Fomichev <sdf@google.com>
    Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>

Signed-off-by: Jiri Benc <jbenc@redhat.com>