fpu/types.h: KABI fix Part I
Bugzilla: http://bugzilla.redhat.com/1837321

Upstream Status: RHEL8 only

I have chosen to break out this KABI workaround due to its signficance in
affecting the rest of this changeset.

The previous commit breaks KABI in struct fpu (specifically,
task_struct->thread->fpu).  The size of struct fpu currently cannot be
changed due to the BUILD_BUG_ON() check compare the sizes and offsets of
fpstate and the soon-to-be-removed fpregs_state.

However, the size of struct fpu CAN be changed because the size of the
struct is variable in length.  The BUILD_BUG_ON()'s are eventually removed
in later commits.  Using the standard RH_KABI wrappers here causes the
above BUILD_BUG_ON() to stop the compile, so for now just use the old
__GENKSYMS__ wrapper which will be removed in later commits.

Fixes: d81d76c7ca9 ("x86/fpu: Provide struct fpstate")
Signed-off-by: Prarit Bhargava <prarit@redhat.com>