perf buildid-cache: Add test for PE executable
Bugzilla: https://bugzilla.redhat.com/2003981

upstream
========
commit 34968b9327c83589e2867af3c5b9fd993666a514
Author: Nicholas Fraser <nfraser@codeweavers.com>
Date: Wed Feb 24 14:59:16 2021 -0500

description
===========
This builds on the previous changes to tests/shell/buildid.sh, adding
tests for a PE file. It adds it to the build-id cache manually and, if
Wine is available, runs it under "perf record" and verifies that it was
added automatically.

If wine is not installed, only warnings are printed; the test can still
exit 0.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>