perf cs-etm: Update ETM metadata format
Bugzilla: https://bugzilla.redhat.com/2003981

upstream
========
commit 42b2b570b34afb5fb9dc16ac77cb332194136a85
Author: Mike Leach <mike.leach@linaro.org>
Date: Wed Feb 24 09:48:30 2021 -0700

description
===========
The current fixed metadata version format (version 0), means that adding
metadata parameter items renders files from a previous version of perf
unreadable. Per CPU parameters appear in a fixed order, but there is no
field to indicate the number of ETM parameters per CPU.

This patch updates the per CPU parameter blocks to include a NR_PARAMs
value which indicates the number of parameters in the block.

The header version is incremented to 1. Fixed ordering is retained,
new ETM parameters are added to the end of the list.

The reader code is updated to be able to read current version 0 files,
For version 1, the reader will read the number of parameters in the
per CPU block. This allows the reader to process older or newer files
that may have different numbers of parameters than in use at the
time perf was built.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>