perf cs-etm: Fix bitmap for option
Bugzilla: https://bugzilla.redhat.com/2003981

upstream
========
commit 8c559e8d68630d64d932bada633705f6551427df
Author: Suzuki K Poulose <suzuki.poulose@arm.com>
Date: Wed Feb 24 09:48:32 2021 -0700

description
===========
When set option with macros ETM_OPT_CTXTID and ETM_OPT_TS, it wrongly
takes these two values (14 and 28 prespectively) as bit masks, but
actually both are the offset for bits.  But this doesn't lead to further
failure due to the AND logic operation will be always true for
ETM_OPT_CTXTID / ETM_OPT_TS.

This patch uses the BIT() macro for option bits, thus it can request the
correct bitmaps for "contextid" and "timestamp" when calling
cs_etm_set_option().

[Extract the change as a separate patch for easier review]

Conflicts:
==========
Previously accepted 6fc5baf54717 ("perf cs-etm: Fix bitmap for option")
(which got to v5.12) was basically reworked by this patch, which is
something like a "v2". Both patches are in upstream kernel repository
and they conflict with each other.

Taking this patch in order to keep the code as synced with upstream as
possible.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>