perf report: Create option to disable raw event ordering
Bugzilla: https://bugzilla.redhat.com/2003981

upstream
========
commit 2e989f82181cd414599c6afbc5a666356a3d1dd1
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Fri Feb 19 15:00:05 2021 +0800

description
===========
Warning "dso not found" is reported when using "perf report -D".

 66702781413407 0x32c0 [0x30]: PERF_RECORD_SAMPLE(IP, 0x2): 28177/28177: 0x55e493e00563 period: 106578 addr: 0
  ... thread: perf:28177
  ...... dso: <not found>

 66702727832429 0x9dd8 [0x38]: PERF_RECORD_COMM exec: triad_loop:28177/28177

The PERF_RECORD_SAMPLE event (timestamp: 66702781413407) should be after the
PERF_RECORD_COMM event (timestamp: 66702727832429), but it's early processed.

So for most of cases, it makes sense to keep the event ordered even for dump
mode. But it would be also useful to disable ordered_events for reporting raw
dump to see events as they are stored in the perf.data file.

So now, set ordered_events by default to true and add a new option
'disable-order' to disable it. For example,

  perf report -D --disable-order

Fixes: 977f739b7126b ("perf report: Disable ordered_events for raw dump")

Signed-off-by: Michael Petlan <mpetlan@redhat.com>