perf test: Add CSV summary test
Bugzilla: https://bugzilla.redhat.com/2003981

upstream
========
commit 0f7ff383937b24a3db72234a37e8b724acda8ad3
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Fri Mar 19 15:01:56 2021 +0800

description
===========
The patch "perf stat: Align CSV output for summary mode" aligned CSV
output and added "summary" to the first column of summary lines.

Now we check if the "summary" string is added to the CSV output.

If we set '--no-csv-summary' option, the "summary" string would not be
added, also check with this case.

Committer testing:

  $ perf test csv
  84: perf stat csv summary test     : Ok
  $

Signed-off-by: Michael Petlan <mpetlan@redhat.com>