perf pmu: Add pmu_events_map__find() function to find the common PMU map for the system
Bugzilla: https://bugzilla.redhat.com/2003981

upstream
========
commit e126bef55f1dfb44440d632f9aae66af3240a435
Author: John Garry <john.garry@huawei.com>
Date: Wed Apr 7 18:32:47 2021 +0800

description
===========
Add a function to find the common PMU map for the system.

For arm64, a special variant is added. This is because arm64 supports
heterogeneous CPU systems. As such, it cannot be guaranteed that the
cpumap is same for all CPUs. So in case of heterogeneous systems, don't
return a cpumap.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>