perf hists: Split hists_stats from events_stats
Bugzilla: https://bugzilla.redhat.com/2003981

upstream
========
commit 0f0abbace3cddc92aaed2db3783c9c501354b3be
Author: Namhyung Kim <namhyung@kernel.org>
Date: Mon Apr 26 18:37:13 2021 -0700

description
===========
Each struct hists have events_stats but most of the fields were not
used.  It's to count number of samples and periods whether filtered or
not.  And other fields are used only by evlist.

So it'd be better to split hists_stats and events_stats to reduce
wasted memory in the struct hists.  This makes the output of event
statistics in the perf report compact by skipping 0 events in each
evsel/hists.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>