perf jit: Let convert_timestamp() to be backwards-compatible
Bugzilla: https://bugzilla.redhat.com/2003981

upstream
========
commit aa616f5a8a2d22a179d5502ebd85045af66fa656
Author: Leo Yan <leo.yan@linaro.org>
Date: Wed Apr 28 20:09:13 2021 +0800

description
===========
Commit d110162cafc80dad ("perf tsc: Support cap_user_time_short for
event TIME_CONV") supports the extended parameters for event TIME_CONV,
but it broke the backwards compatibility, so any perf data file with old
event format fails to convert timestamp.

This patch introduces a helper event_contains() to check if an event
contains a specific member or not.  For the backwards-compatibility, if
the event size confirms the extended parameters are supported in the
event TIME_CONV, then copies these parameters.

Committer notes:

To make this compiler backwards compatible add this patch:

  -       struct perf_tsc_conversion tc = { 0 };
  +       struct perf_tsc_conversion tc = { .time_shift = 0, };

Fixes: d110162cafc8 ("perf tsc: Support cap_user_time_short for event TIME_CONV")

Signed-off-by: Michael Petlan <mpetlan@redhat.com>