perf tools: Add 'cgroup-switches' software event
Bugzilla: https://bugzilla.redhat.com/2003981

upstream
========
commit fb6c79d7261afb7e942251254ea47951c2a9a706
Author: Namhyung Kim <namhyung@kernel.org>
Date: Wed Feb 10 17:33:27 2021 +0900

description
===========
It counts how often cgroups are changed actually during the context
switches.

  # perf stat -a -e context-switches,cgroup-switches -a sleep 1

   Performance counter stats for 'system wide':

              11,267      context-switches
              10,950      cgroup-switches

         1.015634369 seconds time elapsed

Committer notes:

The kernel patches landed in v5.13, but this entry wasn't filled in
perf's parse-events tables, which was leading to a segfault when running
'perf list' on a kernel with that feature, as reported by Thomas
Richter.

Also removed the part touching tools/include/uapi/linux/perf_event.h as
it was updated in the usual sync with the kernel UAPI headers, in a
previous, already upstream, patch.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>