perf scripts python: exported-sql-viewer.py: Fix warning display
Bugzilla: https://bugzilla.redhat.com/2003981

upstream
========
commit f56299a9c998e0bfbd4ab07cafe9eb8444512448
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri May 21 12:20:53 2021 +0300

description
===========
Deprecation warnings are useful only for the developer, not an end user.
Display warnings only when requested using the python -W option. This
stops the display of warnings like:

 tools/perf/scripts/python/exported-sql-viewer.py:5102: DeprecationWarning:
         an integer is required (got type PySide2.QtCore.Qt.AlignmentFlag).
         Implicit conversion to integers using __int__ is deprecated, and
         may be removed in a future version of Python.
    err = app.exec_()

Since the warning can be fixed only in PySide2, we must wait for it to
be finally fixed there.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>