perf stat: Honor event config name on --no-merge
Bugzilla: https://bugzilla.redhat.com/2003981

upstream
========
commit 3cc84399e9b60463bc39cf352ffd8bccb92e02bd
Author: Namhyung Kim <namhyung@kernel.org>
Date: Wed Jun 2 14:22:41 2021 -0700

description
===========
If user gave an event name explicitly, it should be displayed in the
output as is.  But with --no-merge option it adds a pmu name at the
end so might confuse users.

Actually this is true for hybrid pmus, I think we should do the same
for others.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>