perf metricgroup: Return error code from metricgroup__add_metric_sys_event_iter()
Bugzilla: https://bugzilla.redhat.com/2003981

upstream
========
commit fe7a98b9d9b36e5c8a22d76b67d29721f153f66e
Author: John Garry <john.garry@huawei.com>
Date: Thu Jun 10 22:33:00 2021 +0800

description
===========
The error code is not set at all in the sys event iter function.

This may lead to an uninitialized value of "ret" in
metricgroup__add_metric() when no CPU metric is added.

Fix by properly setting the error code.

It is not necessary to init "ret" to 0 in metricgroup__add_metric(), as
if we have no CPU or sys event metric matching, then "has_match" should
be 0 and "ret" is set to -EINVAL.

However gcc cannot detect that it may not have been set after the
map_for_each_metric() loop for CPU metrics, which is strange.

Fixes: be335ec28efa8 ("perf metricgroup: Support adding metrics for system PMUs")

Signed-off-by: Michael Petlan <mpetlan@redhat.com>