perf/x86/intel/uncore: Drop unnecessary NULL checks after container_of()
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit 440e906702410f59ae5397ec9e3b639edb53f80e
Author: Guenter Roeck <linux@roeck-us.net>
Date: Mon May 10 15:48:49 2021 -0700

description
===========
The parameter passed to the pmu_enable() and pmu_disable() functions can not be
NULL because it is dereferenced by the caller.

That means the result of container_of() on that parameter can also never be NULL.
The existing NULL checks are therefore unnecessary and misleading. Remove them.

This change was made automatically with the following Coccinelle script.

  @@
  type t;
  identifier v;
  statement s;
  @@

  <+...
  (
    t v = container_of(...);
  |
    v = container_of(...);
  )
    ...
    when != v
  - if (\( !v \| v == NULL \) ) s
  ...+>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>