perf: arm-ccn: Use scnprintf() for robustness
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit 06236821aeac480a0835dd8dd9fb20e3b5a5d80d
Author: Takashi Iwai <tiwai@suse.de>
Date: Sun Mar 15 10:37:15 2020 +0100

description
===========
snprintf() is a hard-to-use function, it's especially difficult to use
it for concatenating substrings in a buffer with a limited size.
Since snprintf() returns the would-be-output size, not the actual
size, the subsequent use of snprintf() may point to the incorrect
position easily.  Although the current code doesn't actually overflow
the buffer, it's an incorrect usage.

This patch replaces such snprintf() calls with a safer version,
scnprintf().

Signed-off-by: Michael Petlan <mpetlan@redhat.com>