perf/x86/intel/uncore: Fix integer overflow on 23 bit left shift of a u32
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit 0b3a8738b76fe2087f7bc2bd59f4c78504c79180
Author: Colin Ian King <colin.king@canonical.com>
Date: Tue Jul 6 12:45:53 2021 +0100

description
===========
The u32 variable pci_dword is being masked with 0x1fffffff and then left
shifted 23 places. The shift is a u32 operation,so a value of 0x200 or
more in pci_dword will overflow the u32 and only the bottow 32 bits
are assigned to addr. I don't believe this was the original intent.
Fix this by casting pci_dword to a resource_size_t to ensure no
overflow occurs.

Note that the mask and 12 bit left shift operation does not need this
because the mask SNR_IMC_MMIO_MEM0_MASK and shift is always a 32 bit
value.

Fixes: ee49532b38dd ("perf/x86/intel/uncore: Add IMC uncore support for Snow Ridge")
Addresses-Coverity: ("Unintentional integer overflow")

Signed-off-by: Michael Petlan <mpetlan@redhat.com>