perf inject: Add facility to do in place update
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit 2a525f6a5502bfd80568e6befb84053cf650ad25
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Apr 30 10:03:01 2021 +0300

description
===========
When there is a need to modify only timestamps, it is much simpler and
quicker to do it to the existing file rather than re-write all the
contents.

In preparation for that, add the ability to modify the input file in place.
In practice that just means making the file descriptor and mmaps writable.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>