perf inject: Add --vm-time-correlation option
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit 83d7f5f1ad0ec6072bb5b4ca32795f0c199424e2
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Apr 30 10:03:02 2021 +0300

description
===========
Intel PT timestamps are affected by virtualization. Add a new option
that will allow the Intel PT decoder to correlate the timestamps and
translate the virtual machine timestamps to host timestamps.

The advantages of making this a separate step, rather than a part of
normal decoding are that it is simpler to implement, and it needs to
be done only once.

This patch adds only the option. Later patches add Intel PT support.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>