perf intel-pt: Add a tree for VMCS information
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit 0fc9d338944254561289e8639ac05ebf72b4b082
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Apr 30 10:03:05 2021 +0300

description
===========
Even when VMX TSC Offset is not changing (during perf record), different
virtual machines can have different TSC Offsets. There is a Virtual Machine
Control Structure (VMCS) for each virtual CPU, the address of which is
reported to Intel PT in the VMCS packet. We do not know which VMCS belongs
to which virtual machine, so use a tree to keep track of VMCS information.
Then the decoder will be able to use the current VMCS value to look up the
current TSC Offset.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>