perf cs-etm: Refactor timestamp variable names
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit aadd6ba409344776fc155451684c728822bf4a24
Author: James Clark <james.clark@arm.com>
Date: Mon May 10 17:32:47 2021 +0300

description
===========
Remove ambiguity in variable names relating to timestamps.

A later commit will save the sample kernel timestamp in one of the etm
structs, so name all elements appropriately to avoid confusion.

This is also removes some ambiguity arising from the fact that the
--timestamp argument to perf record refers to sample kernel timestamps,
and the /timestamp/ event modifier refers to CS timestamps, so the term
is overloaded.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>