perf header: Support HYBRID_TOPOLOGY feature
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit f7d74ce32fc1b9b3cbf58c015009d1f616e60c23
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Fri May 14 20:29:47 2021 +0800

description
===========
It is useful to let the user know about the hybrid topology.

Add the HYBRID_TOPOLOGY feature in header to indicate the core CPUs
and the atom CPUs.

With this patch a perf.data generated on a hybrid platform reports
the hybrid CPU list:

  root@otcpl-adl-s-2:~# perf report --header-only -I
  ...
  # hybrid cpu system:
  # cpu_core cpu list : 0-15
  # cpu_atom cpu list : 16-23

For a perf.data generated on a non-hybrid platform, reports a message
that HYBRID_TOPOLOGY is missing:

  root@kbl-ppc:~# perf report --header-only -I
  ...
  # missing features: TRACING_DATA BRANCH_STACK GROUP_DESC AUXTRACE STAT CLOCKID DIR_FORMAT COMPRESSED CLOCK_DATA HYBRID_TOPOLOGY

Signed-off-by: Michael Petlan <mpetlan@redhat.com>