perf cs-etm: Prevent and warn on underflows during timestamp calculation.
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit c1a6165a639c2d85f06be7dd9c29ce1cfd994775
Author: James Clark <james.clark@arm.com>
Date: Mon May 17 16:17:41 2021 +0300

description
===========
When a zero timestamp is encountered, warn once. This is to make
hardware or configuration issues visible. Also suggest that the issue
can be worked around with the --itrace=Z option.

When an underflow with a non-zero timestamp occurs, warn every time.
This is an unexpected scenario, and with increasing timestamps, it's
unlikely that it would occur more than once, therefore it should be
ok to warn every time.

Only try to calculate the timestamp by subtracting the instruction
count if neither of the above cases are true. This makes attempting
to decode files with zero timestamps in non-timeless mode
more consistent. Currently it can half work if the timestamp wraps
around and becomes non-zero, although the behavior is undefined and
unpredictable.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>