perf arm-spe: Correct sample flags for SPE event
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit 2f021954952f105e32383103cac06a095d9a1db6
Author: Leo Yan <leo.yan@linaro.org>
Date: Wed May 19 12:15:43 2021 +0800

description
===========
Now it's hard code to set sample flags for CPU, TIME and TID for SPE
event, which is pointless.

The CPU is useful for sampling only for per-mmap case, it is used to
indicate the AUX trace is associated to which CPU.

The TIME sample is not needed for AUX event, since the time for AUX
event is not really used and this time is a different thing from the
timestamp in Arm SPE trace, the timestamp tracing which is controlled
by Arm SPE's config bit.

The TID sample is not useful for AUX event.

This patch corrects the sample flags for SPE event, it only set CPU
sample bit for per-cpu mmap case.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>