perf arm-spe: Correct sample flags for dummy event
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit e582badf1706a005359ef3e184fa97384e93f83a
Author: Leo Yan <leo.yan@linaro.org>
Date: Wed May 19 12:15:44 2021 +0800

description
===========
The dummy event is mainly used for mmap, the TIME sample is only needed
for per-cpu case so that the perf tool can rely on the correct timing
for parsing symbols.  And the CPU sample is useless for mmap.

The BRANCH_STACK sample bit will be always reset for the dummy event in
the function evsel__config(), so don't need to repeatedly reset it for
Arm SPE specific.

So this patch only enables TIME sample for per-cpu mmap.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>