perf test: Iterate over shell tests in alphabetical order
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit da963834fe6975a11f88d7ad9e9280938476c8dc
Author: Riccardo Mancini <rickyman7@gmail.com>
Date: Wed May 26 01:05:17 2021 +0200

description
===========
The for_each_shell_test macro iterated over all shell tests in the
directory using readdir, which does not guarantee any ordering, causing
problems on certain fs. However, the order in which they are visited
determines the id of the test, in case one wants to run a single test.

This patch replaces readdir with scandir using alphabetical sorting.
This guarantees that, given the same set of tests, all machines will
see the tests in the same order, and, thus, that test ids are
consistent.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>