perf tools: Check mem-loads auxiliary event
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit ddc11da5eb37e27a4b66cddcaf11233ef51b3a79
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Thu May 27 08:16:03 2021 +0800

description
===========
For some platforms, an auxiliary event has to be enabled
simultaneously with the load latency event.

For Alderlake, the auxiliary event is created in "cpu_core" pmu.

So first we need to check the existing of "cpu_core" pmu
and then check if this pmu has auxiliary event.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>