perf mem: Disable 'mem-loads-aux' group before reporting
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit d5a8bd0fcd069819aa48f5e38548e07d5eb3e651
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Thu May 27 08:16:09 2021 +0800

description
===========
For some platforms, such as Alderlake, the 'mem-loads' event is required
to use together with 'mem-loads-aux' within a group and 'mem-loads-aux'
must be the group leader. Now we disable this group before reporting
because 'mem-loads-aux' is just an auxiliary event. It doesn't carry
any valid memory load result. If we show the 'mem-loads-aux' +
'mem-loads' as a group in report, it needs many of changes but they
are totally unnecessary.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>