perf mem: Fix wrong verbose output for recording events
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit a6d9de8427584553b71492071f6ffd7c92ec8b99
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Thu May 27 08:16:08 2021 +0800

description
===========
Current code:

for (j = 0; j < argc; j++, i++)
        rec_argv[i] = argv[j];

if (verbose > 0) {
        pr_debug("calling: record ");

        while (rec_argv[j]) {
                pr_debug("%s ", rec_argv[j]);
                j++;
        }
        pr_debug("\n");
}

The entries of argv[] are copied to the end of rec_argv[], not
copied to the beginning of rec_argv[]. So the index j at
rec_argv[] doesn't point to the first event.

Now we record the start index and end index for events in rec_argv[],
and print them if verbose is enabled.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>