perf probe: Support probes on init functions for offline kernel
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit f338de22192085a476dd1bc9eac5e074a48f34f9
Author: Masami Hiramatsu <mhiramat@kernel.org>
Date: Sat Jun 5 01:28:23 2021 +0900

description
===========
'perf probe' internally checks the probe target is in the text area in
post-process (after analyzing debuginfo). But it fails if the probe
target is in the "inittext".

This is a good limitation for the online kernel because such functions
have gone after booting. However, for using it for boot-time tracing,
user may want to put a probe on init functions.

This skips the post checking process if the target is offline kenrel so
that user can get the probe definition on the init functions.

Without this patch:

  $ perf probe -k ./build-x86_64/vmlinux -D do_mount_root:10
  Probe point 'do_mount_root:10' not found.
    Error: Failed to add events.

With this patch:

  $ perf probe -k ./build-x86_64/vmlinux -D do_mount_root:10
  p:probe/do_mount_root_L10 mount_block_root+300

Signed-off-by: Michael Petlan <mpetlan@redhat.com>