perf probe: Add --bootconfig to output definition in bootconfig format
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit 45237f9898fce54a8cc6d40f7455291e0e6c4277
Author: Masami Hiramatsu <mhiramat@kernel.org>
Date: Sat Jun 5 01:28:43 2021 +0900

description
===========
Now the boot-time tracing supports kprobes events and that must be
written in bootconfig file in the following format.

  ftrace.event.kprobes.<EVENT_NAME>.probes = <PROBE-DEF>

'perf probe' already supports --definition (-D) action to show probe
definitions, but the format is for tracefs:

  [p|r][:EVENT_NAME] <PROBE-DEF>

This patch adds the --bootconfig option for -D action so that it outputs
the probe definitions in bootconfig format. E.g.

  $ perf probe --bootconfig -D "path_lookupat:7 err:s32 s:string"
  ftrace.event.kprobes.path_lookupat_L7.probe = 'path_lookupat.isra.0+309 err_s32=%ax:s32 s_string=+0(%r13):string'

Signed-off-by: Michael Petlan <mpetlan@redhat.com>