perf arm-spe: Bail out if the trace is later than perf event
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit afb5e9e47faf53e0f557e22979076dc1a94ef3d7
Author: Leo Yan <leo.yan@linaro.org>
Date: Wed May 19 15:19:38 2021 +0800

description
===========
It's possible that record in Arm SPE trace is later than perf event and
vice versa.  This asks to correlate the perf events and Arm SPE
synthesized events to be processed in the manner of correct timing.

To achieve the time ordering, this patch reverses the flow, it firstly
calls arm_spe_sample() and then calls arm_spe_decode().  By comparing
the timestamp value and detect the perf event is coming earlier than Arm
SPE trace data, it bails out from the decoding loop, the last record is
pushed into auxtrace stack and is deferred to generate sample.  To track
the timestamp, everytime it updates timestamp for the latest record.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>