perf symbol-elf: Decode dynsym even if symtab exists
Bugzilla: https://bugzilla.redhat.com/2003982

upstream
========
commit 87704345cc602a845be713abdc679d65dc600431
Author: Masami Hiramatsu <mhiramat@kernel.org>
Date: Sun Jul 4 00:35:27 2021 +0900

description
===========
In Fedora34, libc-2.33.so has both .dynsym and .symtab sections and
most of (not all) symbols moved to .dynsym. In this case, perf only
decode the symbols in .symtab, and perf probe can not list up the
functions in the library.

To fix this issue, decode both .symtab and .dynsym sections.

Without this fix,
  -----
  $ ./perf probe -x /usr/lib64/libc-2.33.so -F
  @plt
  @plt
  calloc@plt
  free@plt
  malloc@plt
  memalign@plt
  realloc@plt
  -----

With this fix.

  -----
  $ ./perf probe -x /usr/lib64/libc-2.33.so -F
  @plt
  @plt
  a64l
  abort
  abs
  accept
  accept4
  access
  acct
  addmntent
  -----

Signed-off-by: Michael Petlan <mpetlan@redhat.com>